Refactor all API custom hook code to useAPI folder - Misc changes
