+++ src/components/ui/Alerts/ConnectedAlerts/ConnectedCustomerDeleteAlertDialog.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DeleteAlertDialog } from '../../..';
-import { useDeleteCustomer } from '../../../../hooks/useFetchData/useCustomers';
+import { useDeleteCustomer } from '../../../../hooks/useAPI/useCustomers';
 
 const ConnectedCustomerDeleteAlertDialog = (props) => {
   const { toast, itemNumber, isOpen, onClose, header, body, entityDescription } = props;
