+++ src/hooks/useAPI/useQuoteRequests.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import supabase from '../utils/supabaseClient';
+import supabase from '../../utils/supabaseClient';
 
 export const useQuoteRequests = () => {
   const [quoteRequests, setQuoteRequests] = useState([]);
