+++ src/hooks/useAPI/useQuoteStatuses.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import supabase from '../utils/supabaseClient';
+import supabase from '../../utils/supabaseClient';
 
 export const useQuoteStatuses = () => {
   const [quoteStatuses, setQuoteStatuses] = useState([]);
