+++ src/pages/Customers/CustomerDetails.jsx
@@ -37,13 +37,13 @@ import { MdKeyboardArrowLeft } from 'react-icons/md';
 import { FiArrowRight, FiFileText } from 'react-icons/fi';
 import { TbRuler } from 'react-icons/tb';
 import { formatDate, formatMoneyValue, formatNumber } from '../../utils';
-import { useCustomerTypes } from '../../hooks/useFetchData/useCustomerTypes';
+import { useCustomerTypes } from '../../hooks/useAPI/useCustomerTypes';
 import {
   useFetchCustomerByID,
   useFetchCustomerInvoices,
   useFetchCustomerQuotes,
   useUpdateCustomer
-} from '../../hooks/useFetchData/useCustomers';
+} from '../../hooks/useAPI/useCustomers';
 
 const CustomerById = () => {
   // React Hooks
